Added AbstractSSLContextFactory
So that both BasicSSLContextFactory and OpenSSLContextFactory can share a lot of the
same logic.
